more build/config cleanup
git-svn-id: https://jna.dev.java.net/svn/jna/branches/win64@681 05a1902d-0e30-0410-9027-c8643ccd90df

+++ jnalib/native/dispatch.c
@@ -1727,7 +1727,7 @@ Java_com_sun_jna_Native_getWindowHandle0(JNIEnv *env, jclass classp, jobject w) 
         if (java_home != NULL) {
           if ((prop = newWideCString(env, java_home)) != NULL) {
             const wchar_t* suffix = L"/bin/jawt.dll";
-            int len = wcslen(prop) + wcslen(suffix) + 1;
+            size_t len = wcslen(prop) + wcslen(suffix) + 1;
             path = (wchar_t*)alloca(len * sizeof(wchar_t));
 #ifdef _MSC_VER
             swprintf(path, len, L"%s%s", prop, suffix);
